keyNavigation: split down the item strategy into item and linearItem
