+++ src/plugins/content/accessibility/keyNavigation/strategies/index.js
@@ -22,3 +22,4 @@ export { default as navigatorNavigationStrategy } from './navigatorNavigation';
 export { default as pageNavigationStrategy } from './pageNavigation';
 export { default as rubricsNavigationStrategy } from './rubricsNavigation';
 export { default as itemNavigationStrategy } from './itemNavigation';
+export { default as linearItemNavigationStrategy } from './linearItemNavigation';
